Adding a tool that can copy data from one store to another
The tool also allows transforming the messages in the store during copy.
The possible uses of this tool will be when messages in a store need
data format changes (for e.g. interconversion b/w segmented and non-
segmented logs or if more data needs to be added to the message).